+++ package.json
@@ -28,7 +28,6 @@
     "seed": "ts-node prisma/seed.ts"
   },
   "dependencies": {
-    "@chakra-ui/icons": "^1.1.1",
     "@chakra-ui/react": "^1.8.1",
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.6.0",
