+++ pages/index.tsx
@@ -1,20 +1,16 @@
-import { EditIcon } from "@chakra-ui/icons";
-import { Button, useDisclosure } from "@chakra-ui/react";
-import ImageUploadModal from "@components/ImageUpload/ImageUploadModal";
 import Layout from "@components/Layout";
-import pageTile from "@lib/helpers/frontend/page-title";
+import pageTitle from "@lib/helpers/frontend/page-title";
 import { useUserQuery } from "@lib/hooks/user";
 import type { NextPage } from "next";
 import Head from "next/head";
 
 const Home: NextPage = () => {
   const { user } = useUserQuery();
-  const { isOpen, onClose, onOpen } = useDisclosure();
 
   return (
     <>
       <Head>
-        <title>{pageTile()}</title>
+        <title>{pageTitle()}</title>
         <meta name="description" content="Generated by create next app" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
@@ -24,21 +20,6 @@ const Home: NextPage = () => {
             {JSON.stringify(user, null, 2)}
           </pre>
         )}
-        <Button
-          mt="4"
-          variant="link"
-          onClick={() => onOpen()}
-          leftIcon={<EditIcon />}
-        >
-          Edit image
-        </Button>
-        <ImageUploadModal
-          aspectRatio={3 / 4}
-          isOpen={isOpen}
-          onClose={onClose}
-          shape={"rect"}
-          onUpload={() => new Promise((resolve) => setTimeout(resolve, 500))}
-        />
       </Layout>
     </>
   );
